backend/requirements.txt
flask
psycopg2-binary
python-dotenv
flask-cors
py-algorand-sdk
pyteal
